chore(ci.yml): update versions and actions in CI workflow
The CI workflow has been updated to use newer versions of the PostgreSQL
and NATS containers. The PostgreSQL container now uses the
postgres:14-alpine image, and the NATS container now uses the
nats:2.9.17-alpine image. This ensures that the CI environment is up to
date with the latest versions of these dependencies.
The CI workflow has been updated to use newer versions of the
actions/setup-go and actions/checkout actions. The actions/setup-go
action now uses version 4, and the actions/checkout action now uses
version 3. This ensures that the Go environment is set up correctly and
the code is checked out properly during the CI process.
The CI workflow has been updated to use a newer version of the
codecov/codecov-action action. The action now uses version 3. This
ensures that the coverage report is uploaded correctly to Codecov.
The CI workflow has been updated to use the
dictyBase/workflows/docker-build-push.yaml action. This action is used
to build and push Docker images. The action is inherited from the
dictyBase/workflows repository and is configured to use the current
repository and branch. This ensures that the Docker images are built and
pushed correctly during the CI process.